requirements.txt
svg-turtle==0.4.1
Pillow==9.2.0
cairosvg==2.5.2
numpy==1.23.0
